test: separate articulation ordering coverage
