+++ source/isaaclab/test/assets/test_articulation_ordering.py
@@ -71,24 +71,10 @@ asset_base_stub.AssetBase = _AssetBase
 _inserted_asset_base_stub = "isaaclab.assets.asset_base" not in sys.modules
 sys.modules.setdefault("isaaclab.assets.asset_base", asset_base_stub)
 
-from isaaclab_newton.actuators.kernels import sync_torque_telemetry
 
 from isaaclab.assets.articulation.articulation_cfg import ArticulationCfg
 from isaaclab.assets.articulation.base_articulation import BaseArticulation
 from isaaclab.assets.articulation.base_articulation_data import BaseArticulationData
-from isaaclab.assets.articulation.ordering_kernels import (
-    reorder_2d_backend_to_user,
-    reorder_2d_user_to_backend,
-    reorder_3d_backend_to_user,
-    write_2d_float_user_to_backend_with_indices,
-    write_2d_float_user_to_backend_with_mask,
-    write_joint_state_user_to_backend_with_indices,
-    write_joint_state_user_to_backend_with_mask,
-    write_joint_vel_user_to_backend_with_indices,
-    write_joint_vel_user_to_backend_with_mask,
-    write_scalar_user_to_backend_with_indices,
-    write_scalar_user_to_backend_with_mask,
-)
 
 if _inserted_sim_stub:
     sys.modules.pop("isaaclab.sim", None)
@@ -1175,7 +1161,6 @@ def _make_ordering_resolution_articulation(
         data=data,
         backend_joint_names=["joint"],
         backend_body_names=list(backend_body_names),
-        _mjwarp_body_names=("foot", "base"),
         is_fixed_base=is_fixed_base,
         device="cpu",
     )
@@ -1197,8 +1182,13 @@ def test_base_articulation_ordering_contract_preserves_legacy_subclasses() -> No
 
 
 @pytest.mark.parametrize("ordering", [("foot", "base"), ArticulationOrderingConvention.MJWARP])
-def test_fixed_base_body_ordering_rejects_root_relocation(ordering) -> None:
+def test_fixed_base_body_ordering_rejects_root_relocation(ordering, monkeypatch: pytest.MonkeyPatch) -> None:
     """Reject explicit and symbolic fixed-base orders that move the root from public index zero."""
+    monkeypatch.setattr(
+        ordering_resolvers,
+        "_get_mjwarp_names_from_newton_usd_builder",
+        lambda _: {"joint": ("joint",), "body": ("foot", "base")},
+    )
     articulation = _make_ordering_resolution_articulation(body_ordering=ordering, is_fixed_base=True)
 
     with pytest.raises(ValueError) as exc_info:
@@ -1301,521 +1291,4 @@ def test_build_articulation_name_map_rejects_incomplete_permutation() -> None:
             backend_names=("hip", "knee"),
             user_names=("hip", "wheel"),
             device="cpu",
-        )
-
-
-def test_reorder_2d_backend_to_user_gathers_user_axis() -> None:
-    """Reorder a 2-D backend buffer into public user order."""
-    backend_data = wp.array(
-        np.asarray(
-            [
-                [10.0, 11.0, 12.0],
-                [20.0, 21.0, 22.0],
-            ],
-            dtype=np.float32,
-        ),
-        dtype=wp.float32,
-        device="cpu",
-    )
-    user_to_backend = wp.array(np.asarray([2, 0, 1], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_data = wp.zeros_like(backend_data)
-
-    wp.launch(
-        reorder_2d_backend_to_user,
-        dim=user_data.shape,
-        inputs=[backend_data, user_to_backend],
-        outputs=[user_data],
-        device="cpu",
-    )
-
-    expected = np.asarray([[12.0, 10.0, 11.0], [22.0, 20.0, 21.0]], dtype=np.float32)
-    np.testing.assert_allclose(user_data.numpy(), expected)
-
-
-def test_sync_torque_telemetry_reads_backend_effort_buffers_in_user_order() -> None:
-    """Report torque telemetry in public joint order from backend-order effort buffers."""
-    joint_pos = wp.zeros((1, 3), dtype=wp.float32, device="cpu")
-    joint_vel = wp.zeros_like(joint_pos)
-    joint_pos_target = wp.zeros_like(joint_pos)
-    joint_vel_target = wp.zeros_like(joint_pos)
-    joint_stiffness = wp.zeros_like(joint_pos)
-    joint_damping = wp.zeros_like(joint_pos)
-    effort_limit = wp.full((1, 3), 1000.0, dtype=wp.float32, device="cpu")
-    joint_modes = wp.array(np.asarray([0, 1, 0], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_to_backend = wp.array(np.asarray([2, 0, 1], dtype=np.int32), dtype=wp.int32, device="cpu")
-    sim_bind_joint_effort = wp.array(
-        np.asarray([[100.0, 200.0, 300.0]], dtype=np.float32),
-        dtype=wp.float32,
-        device="cpu",
-    )
-    actuator_computed_effort = wp.array(
-        np.asarray([[10.0, 20.0, 30.0]], dtype=np.float32),
-        dtype=wp.float32,
-        device="cpu",
-    )
-    computed = wp.zeros_like(joint_pos)
-    applied = wp.zeros_like(joint_pos)
-
-    wp.launch(
-        sync_torque_telemetry,
-        dim=joint_pos.shape,
-        inputs=[
-            joint_pos,
-            joint_vel,
-            joint_pos_target,
-            joint_vel_target,
-            joint_stiffness,
-            joint_damping,
-            effort_limit,
-            joint_modes,
-            sim_bind_joint_effort,
-            actuator_computed_effort,
-            user_to_backend,
-            True,
-        ],
-        outputs=[computed, applied],
-        device="cpu",
-    )
-
-    np.testing.assert_allclose(computed.numpy(), np.asarray([[30.0, 100.0, 20.0]], dtype=np.float32))
-    np.testing.assert_allclose(applied.numpy(), np.asarray([[300.0, 100.0, 200.0]], dtype=np.float32))
-
-
-def test_sync_torque_telemetry_keeps_user_order_effort_buffers_unmapped() -> None:
-    """Report torque telemetry directly from user-order actuator buffers."""
-    joint_pos = wp.zeros((1, 3), dtype=wp.float32, device="cpu")
-    joint_modes = wp.array(np.asarray([0, 1, 0], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_to_backend = wp.array(np.asarray([2, 0, 1], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_effort = wp.array(np.asarray([[100.0, 200.0, 300.0]], dtype=np.float32), dtype=wp.float32, device="cpu")
-    user_computed_effort = wp.array(np.asarray([[10.0, 20.0, 30.0]], dtype=np.float32), dtype=wp.float32, device="cpu")
-    computed = wp.zeros_like(joint_pos)
-    applied = wp.zeros_like(joint_pos)
-
-    wp.launch(
-        sync_torque_telemetry,
-        dim=joint_pos.shape,
-        inputs=[
-            joint_pos,
-            wp.zeros_like(joint_pos),
-            wp.zeros_like(joint_pos),
-            wp.zeros_like(joint_pos),
-            wp.zeros_like(joint_pos),
-            wp.zeros_like(joint_pos),
-            wp.full((1, 3), 1000.0, dtype=wp.float32, device="cpu"),
-            joint_modes,
-            user_effort,
-            user_computed_effort,
-            user_to_backend,
-            False,
-        ],
-        outputs=[computed, applied],
-        device="cpu",
-    )
-
-    np.testing.assert_allclose(computed.numpy(), np.asarray([[10.0, 200.0, 30.0]], dtype=np.float32))
-    np.testing.assert_allclose(applied.numpy(), np.asarray([[100.0, 200.0, 300.0]], dtype=np.float32))
-
-
-def test_reorder_2d_user_to_backend_scatters_backend_axis() -> None:
-    """Reorder a 2-D public user buffer into backend order."""
-    user_data = wp.array(
-        np.asarray(
-            [
-                [12.0, 10.0, 11.0],
-                [22.0, 20.0, 21.0],
-            ],
-            dtype=np.float32,
-        ),
-        dtype=wp.float32,
-        device="cpu",
-    )
-    backend_to_user = wp.array(np.asarray([1, 2, 0], dtype=np.int32), dtype=wp.int32, device="cpu")
-    backend_data = wp.zeros_like(user_data)
-
-    wp.launch(
-        reorder_2d_user_to_backend,
-        dim=backend_data.shape,
-        inputs=[user_data, backend_to_user],
-        outputs=[backend_data],
-        device="cpu",
-    )
-
-    expected = np.asarray([[10.0, 11.0, 12.0], [20.0, 21.0, 22.0]], dtype=np.float32)
-    np.testing.assert_allclose(backend_data.numpy(), expected)
-
-
-def test_reorder_3d_backend_to_user_gathers_user_axis() -> None:
-    """Reorder a 3-D backend buffer whose second axis is the public user axis."""
-    backend_data_np = np.asarray(
-        [
-            [[100.0, 101.0], [110.0, 111.0], [120.0, 121.0]],
-            [[200.0, 201.0], [210.0, 211.0], [220.0, 221.0]],
-        ],
-        dtype=np.float32,
-    )
-    backend_data = wp.array(backend_data_np, dtype=wp.float32, device="cpu")
-    user_to_backend = wp.array(np.asarray([1, 2, 0], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_data = wp.zeros_like(backend_data)
-
-    wp.launch(
-        reorder_3d_backend_to_user,
-        dim=user_data.shape,
-        inputs=[backend_data, user_to_backend],
-        outputs=[user_data],
-        device="cpu",
-    )
-
-    np.testing.assert_allclose(user_data.numpy(), backend_data_np[:, [1, 2, 0], :])
-
-
-@pytest.mark.parametrize("selection", ["indices", "mask"])
-@pytest.mark.parametrize("writer", ["scalar", "float", "velocity", "state"])
-def test_fused_identity_writer_supports_aliased_outputs(selection: str, writer: str) -> None:
-    """Preserve identity-order writes when public and backend outputs alias."""
-    user_to_backend = wp.array(np.asarray([0, 1, 2], dtype=np.int32), dtype=wp.int32, device="cpu")
-    env_ids = wp.array(np.asarray([0], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_ids = wp.array(np.asarray([1], dtype=np.int32), dtype=wp.int32, device="cpu")
-    env_mask = wp.array(np.asarray([True, False]), dtype=wp.bool, device="cpu")
-    user_mask = wp.array(np.asarray([False, True, False]), dtype=wp.bool, device="cpu")
-    initial = np.asarray([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]], dtype=np.float32)
-    input_data = wp.array(
-        np.asarray([[101.0, 102.0, 103.0], [104.0, 105.0, 106.0]], dtype=np.float32),
-        dtype=wp.float32,
-        device="cpu",
-    )
-
-    def expected_with_selected_value(source: np.ndarray, value: float) -> np.ndarray:
-        expected = source.copy()
-        expected[0, 1] = value
-        return expected
-
-    if writer == "scalar":
-        data = wp.array(initial, dtype=wp.float32, device="cpu")
-        if selection == "indices":
-            wp.launch(
-                write_scalar_user_to_backend_with_indices,
-                dim=(env_ids.shape[0], user_ids.shape[0]),
-                inputs=[42.0, env_ids, user_ids, user_to_backend, False],
-                outputs=[data, data],
-                device="cpu",
-            )
-        else:
-            wp.launch(
-                write_scalar_user_to_backend_with_mask,
-                dim=data.shape,
-                inputs=[42.0, env_mask, user_mask, user_to_backend, False],
-                outputs=[data, data],
-                device="cpu",
-            )
-        np.testing.assert_allclose(data.numpy(), expected_with_selected_value(initial, 42.0))
-        return
-
-    if writer == "float":
-        data = wp.array(initial, dtype=wp.float32, device="cpu")
-        if selection == "indices":
-            wp.launch(
-                write_2d_float_user_to_backend_with_indices,
-                dim=(env_ids.shape[0], user_ids.shape[0]),
-                inputs=[input_data, env_ids, user_ids, user_to_backend, False, True],
-                outputs=[data, data],
-                device="cpu",
-            )
-        else:
-            wp.launch(
-                write_2d_float_user_to_backend_with_mask,
-                dim=data.shape,
-                inputs=[input_data, env_mask, user_mask, user_to_backend, False],
-                outputs=[data, data],
-                device="cpu",
-            )
-        np.testing.assert_allclose(data.numpy(), expected_with_selected_value(initial, 102.0))
-        return
-
-    initial_velocity = initial + 10.0
-    initial_previous_velocity = initial + 20.0
-    initial_acceleration = initial + 30.0
-    velocity = wp.array(initial_velocity, dtype=wp.float32, device="cpu")
-    previous_velocity = wp.array(initial_previous_velocity, dtype=wp.float32, device="cpu")
-    acceleration = wp.array(initial_acceleration, dtype=wp.float32, device="cpu")
-
-    if writer == "velocity":
-        if selection == "indices":
-            wp.launch(
-                write_joint_vel_user_to_backend_with_indices,
-                dim=(env_ids.shape[0], user_ids.shape[0]),
-                inputs=[input_data, env_ids, user_ids, user_to_backend, False, True],
-                outputs=[velocity, previous_velocity, acceleration, velocity],
-                device="cpu",
-            )
-        else:
-            wp.launch(
-                write_joint_vel_user_to_backend_with_mask,
-                dim=velocity.shape,
-                inputs=[input_data, env_mask, user_mask, user_to_backend, False],
-                outputs=[velocity, previous_velocity, acceleration, velocity],
-                device="cpu",
-            )
-        np.testing.assert_allclose(velocity.numpy(), expected_with_selected_value(initial_velocity, 102.0))
-        np.testing.assert_allclose(
-            previous_velocity.numpy(), expected_with_selected_value(initial_previous_velocity, 102.0)
-        )
-        np.testing.assert_allclose(acceleration.numpy(), expected_with_selected_value(initial_acceleration, 0.0))
-        return
-
-    position_data = wp.array(
-        np.asarray([[201.0, 202.0, 203.0], [204.0, 205.0, 206.0]], dtype=np.float32),
-        dtype=wp.float32,
-        device="cpu",
-    )
-    velocity_data = wp.array(
-        np.asarray([[301.0, 302.0, 303.0], [304.0, 305.0, 306.0]], dtype=np.float32),
-        dtype=wp.float32,
-        device="cpu",
-    )
-    initial_position = initial + 40.0
-    position = wp.array(initial_position, dtype=wp.float32, device="cpu")
-    if selection == "indices":
-        wp.launch(
-            write_joint_state_user_to_backend_with_indices,
-            dim=(env_ids.shape[0], user_ids.shape[0]),
-            inputs=[position_data, velocity_data, env_ids, user_ids, user_to_backend, False, True],
-            outputs=[position, velocity, previous_velocity, acceleration, position, velocity],
-            device="cpu",
-        )
-    else:
-        wp.launch(
-            write_joint_state_user_to_backend_with_mask,
-            dim=position.shape,
-            inputs=[position_data, velocity_data, env_mask, user_mask, user_to_backend, False],
-            outputs=[position, velocity, previous_velocity, acceleration, position, velocity],
-            device="cpu",
-        )
-    np.testing.assert_allclose(position.numpy(), expected_with_selected_value(initial_position, 202.0))
-    np.testing.assert_allclose(velocity.numpy(), expected_with_selected_value(initial_velocity, 302.0))
-    np.testing.assert_allclose(
-        previous_velocity.numpy(), expected_with_selected_value(initial_previous_velocity, 302.0)
-    )
-    np.testing.assert_allclose(acceleration.numpy(), expected_with_selected_value(initial_acceleration, 0.0))
-
-
-def test_write_scalar_user_to_backend_with_indices_updates_user_and_backend_buffers() -> None:
-    """Fuse indexed scalar writes into user and backend-order buffers."""
-    env_ids = wp.array(np.asarray([0, 2], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_ids = wp.array(np.asarray([2, 0], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_to_backend = wp.array(np.asarray([1, 2, 0], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_data = wp.zeros((3, 3), dtype=wp.float32, device="cpu")
-    backend_data = wp.zeros((3, 3), dtype=wp.float32, device="cpu")
-
-    wp.launch(
-        write_scalar_user_to_backend_with_indices,
-        dim=(env_ids.shape[0], user_ids.shape[0]),
-        inputs=[4.5, env_ids, user_ids, user_to_backend, True],
-        outputs=[user_data, backend_data],
-        device="cpu",
-    )
-
-    expected_user = np.asarray([[4.5, 0.0, 4.5], [0.0, 0.0, 0.0], [4.5, 0.0, 4.5]], dtype=np.float32)
-    expected_backend = np.asarray([[4.5, 4.5, 0.0], [0.0, 0.0, 0.0], [4.5, 4.5, 0.0]], dtype=np.float32)
-    np.testing.assert_allclose(user_data.numpy(), expected_user)
-    np.testing.assert_allclose(backend_data.numpy(), expected_backend)
-
-
-def test_write_scalar_user_to_backend_with_mask_updates_selected_entries() -> None:
-    """Fuse masked scalar writes into user and backend-order buffers."""
-    env_mask = wp.array(np.asarray([True, False], dtype=bool), dtype=wp.bool, device="cpu")
-    user_mask = wp.array(np.asarray([False, True, True], dtype=bool), dtype=wp.bool, device="cpu")
-    user_to_backend = wp.array(np.asarray([1, 2, 0], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_data = wp.zeros((2, 3), dtype=wp.float32, device="cpu")
-    backend_data = wp.zeros((2, 3), dtype=wp.float32, device="cpu")
-
-    wp.launch(
-        write_scalar_user_to_backend_with_mask,
-        dim=user_data.shape,
-        inputs=[7.25, env_mask, user_mask, user_to_backend, True],
-        outputs=[user_data, backend_data],
-        device="cpu",
-    )
-
-    expected_user = np.asarray([[0.0, 7.25, 7.25], [0.0, 0.0, 0.0]], dtype=np.float32)
-    expected_backend = np.asarray([[7.25, 0.0, 7.25], [0.0, 0.0, 0.0]], dtype=np.float32)
-    np.testing.assert_allclose(user_data.numpy(), expected_user)
-    np.testing.assert_allclose(backend_data.numpy(), expected_backend)
-
-
-def test_write_2d_float_user_to_backend_with_indices_updates_user_and_backend_buffers() -> None:
-    """Fuse partial user-order writes into user and backend-order buffers."""
-    input_data = wp.array(
-        np.asarray(
-            [
-                [10.0, 11.0],
-                [20.0, 21.0],
-            ],
-            dtype=np.float32,
-        ),
-        dtype=wp.float32,
-        device="cpu",
-    )
-    env_ids = wp.array(np.asarray([0, 2], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_ids = wp.array(np.asarray([2, 0], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_to_backend = wp.array(np.asarray([1, 2, 0], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_data = wp.zeros((3, 3), dtype=wp.float32, device="cpu")
-    backend_data = wp.zeros((3, 3), dtype=wp.float32, device="cpu")
-
-    wp.launch(
-        write_2d_float_user_to_backend_with_indices,
-        dim=input_data.shape,
-        inputs=[input_data, env_ids, user_ids, user_to_backend, True, False],
-        outputs=[user_data, backend_data],
-        device="cpu",
-    )
-
-    expected_user = np.asarray([[11.0, 0.0, 10.0], [0.0, 0.0, 0.0], [21.0, 0.0, 20.0]], dtype=np.float32)
-    expected_backend = np.asarray([[10.0, 11.0, 0.0], [0.0, 0.0, 0.0], [20.0, 21.0, 0.0]], dtype=np.float32)
-    np.testing.assert_allclose(user_data.numpy(), expected_user)
-    np.testing.assert_allclose(backend_data.numpy(), expected_backend)
-
-
-def test_write_2d_float_user_to_backend_with_mask_updates_selected_entries() -> None:
-    """Fuse masked user-order writes into user and backend-order buffers."""
-    input_data = wp.array(
-        np.asarray(
-            [
-                [10.0, 11.0, 12.0],
-                [20.0, 21.0, 22.0],
-            ],
-            dtype=np.float32,
-        ),
-        dtype=wp.float32,
-        device="cpu",
-    )
-    env_mask = wp.array(np.asarray([True, False], dtype=bool), dtype=wp.bool, device="cpu")
-    user_mask = wp.array(np.asarray([False, True, True], dtype=bool), dtype=wp.bool, device="cpu")
-    user_to_backend = wp.array(np.asarray([1, 2, 0], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_data = wp.zeros_like(input_data)
-    backend_data = wp.zeros_like(input_data)
-
-    wp.launch(
-        write_2d_float_user_to_backend_with_mask,
-        dim=input_data.shape,
-        inputs=[input_data, env_mask, user_mask, user_to_backend, True],
-        outputs=[user_data, backend_data],
-        device="cpu",
-    )
-
-    expected_user = np.asarray([[0.0, 11.0, 12.0], [0.0, 0.0, 0.0]], dtype=np.float32)
-    expected_backend = np.asarray([[12.0, 0.0, 11.0], [0.0, 0.0, 0.0]], dtype=np.float32)
-    np.testing.assert_allclose(user_data.numpy(), expected_user)
-    np.testing.assert_allclose(backend_data.numpy(), expected_backend)
-
-
-def test_write_joint_state_user_to_backend_with_indices_updates_history_and_backend_buffers() -> None:
-    """Fuse indexed joint state writes into user cache, history, and backend buffers."""
-    position = wp.array(np.asarray([[1.0, 2.0], [3.0, 4.0]], dtype=np.float32), dtype=wp.float32, device="cpu")
-    velocity = wp.array(np.asarray([[5.0, 6.0], [7.0, 8.0]], dtype=np.float32), dtype=wp.float32, device="cpu")
-    env_ids = wp.array(np.asarray([0, 2], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_ids = wp.array(np.asarray([2, 0], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_to_backend = wp.array(np.asarray([1, 2, 0], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_pos = wp.zeros((3, 3), dtype=wp.float32, device="cpu")
-    user_vel = wp.zeros((3, 3), dtype=wp.float32, device="cpu")
-    user_prev_vel = wp.zeros((3, 3), dtype=wp.float32, device="cpu")
-    user_acc = wp.ones((3, 3), dtype=wp.float32, device="cpu")
-    backend_pos = wp.zeros((3, 3), dtype=wp.float32, device="cpu")
-    backend_vel = wp.zeros((3, 3), dtype=wp.float32, device="cpu")
-
-    wp.launch(
-        write_joint_state_user_to_backend_with_indices,
-        dim=position.shape,
-        inputs=[position, velocity, env_ids, user_ids, user_to_backend, True, False],
-        outputs=[user_pos, user_vel, user_prev_vel, user_acc, backend_pos, backend_vel],
-        device="cpu",
-    )
-
-    expected_user_pos = np.asarray([[2.0, 0.0, 1.0], [0.0, 0.0, 0.0], [4.0, 0.0, 3.0]], dtype=np.float32)
-    expected_user_vel = np.asarray([[6.0, 0.0, 5.0], [0.0, 0.0, 0.0], [8.0, 0.0, 7.0]], dtype=np.float32)
-    expected_backend_pos = np.asarray([[1.0, 2.0, 0.0], [0.0, 0.0, 0.0], [3.0, 4.0, 0.0]], dtype=np.float32)
-    expected_backend_vel = np.asarray([[5.0, 6.0, 0.0], [0.0, 0.0, 0.0], [7.0, 8.0, 0.0]], dtype=np.float32)
-    np.testing.assert_allclose(user_pos.numpy(), expected_user_pos)
-    np.testing.assert_allclose(user_vel.numpy(), expected_user_vel)
-    np.testing.assert_allclose(user_prev_vel.numpy(), expected_user_vel)
-    np.testing.assert_allclose(backend_pos.numpy(), expected_backend_pos)
-    np.testing.assert_allclose(backend_vel.numpy(), expected_backend_vel)
-    np.testing.assert_allclose(user_acc.numpy()[[0, 2]][:, [0, 2]], np.zeros((2, 2), dtype=np.float32))
-
-
-def test_write_joint_vel_user_to_backend_with_mask_updates_selected_entries() -> None:
-    """Fuse masked joint velocity writes into user cache, history, and backend buffers."""
-    velocity = wp.array(
-        np.asarray([[10.0, 11.0, 12.0], [20.0, 21.0, 22.0]], dtype=np.float32),
-        dtype=wp.float32,
-        device="cpu",
-    )
-    env_mask = wp.array(np.asarray([True, False], dtype=bool), dtype=wp.bool, device="cpu")
-    user_mask = wp.array(np.asarray([False, True, True], dtype=bool), dtype=wp.bool, device="cpu")
-    user_to_backend = wp.array(np.asarray([1, 2, 0], dtype=np.int32), dtype=wp.int32, device="cpu")
-    user_vel = wp.zeros_like(velocity)
-    user_prev_vel = wp.zeros_like(velocity)
-    user_acc = wp.ones_like(velocity)
-    backend_vel = wp.zeros_like(velocity)
-
-    wp.launch(
-        write_joint_vel_user_to_backend_with_mask,
-        dim=velocity.shape,
-        inputs=[velocity, env_mask, user_mask, user_to_backend, True],
-        outputs=[user_vel, user_prev_vel, user_acc, backend_vel],
-        device="cpu",
-    )
-
-    expected_user = np.asarray([[0.0, 11.0, 12.0], [0.0, 0.0, 0.0]], dtype=np.float32)
-    expected_backend = np.asarray([[12.0, 0.0, 11.0], [0.0, 0.0, 0.0]], dtype=np.float32)
-    np.testing.assert_allclose(user_vel.numpy(), expected_user)
-    np.testing.assert_allclose(user_prev_vel.numpy(), expected_user)
-    np.testing.assert_allclose(backend_vel.numpy(), expected_backend)
-    np.testing.assert_allclose(user_acc.numpy()[0, 1:], np.zeros((2,), dtype=np.float32))
-
-
-def test_newton_actuator_defaults_follow_requested_public_joint_order() -> None:
-    """Convert Newton actuator gain snapshots and managed IDs into public joint order."""
-    from isaaclab_newton.actuators.adapter import build_newton_actuator_defaults
-
-    controller = types.SimpleNamespace(
-        kp=wp.array((10.0, 30.0, 11.0, 31.0), dtype=wp.float32, device="cpu"),
-        kd=wp.array((1.0, 3.0, 1.1, 3.1), dtype=wp.float32, device="cpu"),
-    )
-    actuator = types.SimpleNamespace(
-        controller=controller,
-        indices=wp.array((0, 2, 3, 5), dtype=wp.uint32, device="cpu"),
-    )
-
-    stiffness, damping, managed = build_newton_actuator_defaults(
-        actuators=[actuator],
-        num_envs=2,
-        num_joints=3,
-        dof_offset=0,
-        device="cpu",
-        joint_user_to_backend_indices=(2, 0, 1),
-    )
-
-    torch.testing.assert_close(stiffness, torch.tensor([[30.0, 10.0, 0.0], [31.0, 11.0, 0.0]]))
-    torch.testing.assert_close(damping, torch.tensor([[3.0, 1.0, 0.0], [3.1, 1.1, 0.0]]))
-    torch.testing.assert_close(managed, torch.tensor([0, 1], dtype=torch.int32))
-
-
-def test_newton_actuator_defaults_reject_incomplete_joint_permutation() -> None:
-    """Reject malformed actuator-default ordering maps with an actionable error."""
-    from isaaclab_newton.actuators.adapter import build_newton_actuator_defaults
-
-    with pytest.raises(
-        ValueError,
-        match=(
-            r"joint_user_to_backend_indices must contain each backend joint index exactly once; "
-            r"expected a permutation of 0\.\.2, got \(0, 0, 2\)\."
-        ),
-    ):
-        build_newton_actuator_defaults(
-            actuators=[],
-            num_envs=1,
-            num_joints=3,
-            dof_offset=0,
-            device="cpu",
-            joint_user_to_backend_indices=(0, 0, 2),
         )
